Re-organise browerify transform
Move babelify transform out of browserify key in package.json. This
change is due to an error on deploy when requiring this module from a
project that uses a newer version of babel. This error is due to an
incompatability in the babelrc referencing an old key that isn't used by
babel 6.

Since this project exports the babel-ified version in the dist/ folder,
there isn't any need to run browserify or babelify over the code when using
it as a dependency in another project.

Add npmignore file to slim down production release. This file will ignore
any files not needed to use this module as a dependency.